huge pep8 improvements, commenting and renaming
